Fixed expense bug
Introduced bug from the split from CarbonMonoxide that prevented expenses
from being queried from AtTask. This was due specifically to code that
abstracted filter building for payroll. For expenses to be queried
filter builder must see that it has been constrained by a datatime type.
Unfortunately I was creating a filter constraint with a string encoded
with the datetime, and my filter code isn't smart enough to recognize
that as having been called with a date. The correction was to create
the same filter constraint but this time with the datetime object. Stupid
bug created by questionable design.