Representing Payroll's weekEnding as GMT time
The week ending property from AtTask's API is returned in the format
of the Sunday at 0:00GMT, which gets interpreted as -5/-4 for EST.
This falsely is presented the user as the week ending on Saturday, not
Sunday. The behavior now is to shift the local time to the equivalent time
GMT, presenting the user the appropriate Sunday week ending date.